+++ src/components/Loader/Loader.jsx
@@ -1,6 +1,8 @@
 import * as React from 'react';
 import './Loader.css';
 
+// TODO Add blank/background content to keep footer on bottom
+
 class Loader extends React.Component {
   render() {
     return (
